refactor(api): move figure metadata helpers to common visualization
