+++ src/qdash/common/visualization/__init__.py
@@ -1,5 +1,6 @@
 """Visualization helpers shared across runtime surfaces."""
 
+from qdash.common.visualization.figure_metadata import figure_role_suffix, set_figure_role
 from qdash.common.visualization.metrics_chart import (
     ChipGeometry,
     build_chip_geometry,
@@ -15,5 +16,7 @@ __all__ = [
     "chip_geometry_from_topology",
     "create_data_matrix",
     "create_qubit_heatmap",
+    "figure_role_suffix",
     "get_qubit_position",
+    "set_figure_role",
 ]
