+++ src/qdash/common/visualization/figure_metadata.py
